requirements.txt
numpy
Pillow
tensorflow
opencv-python
matplotlib
scipy
scikit-learn
